+++ certbot/tests/util_test.py
@@ -632,8 +632,8 @@ class AtexitRegisterTest(unittest.TestCase):
             atexit_func(*args[1:], **kwargs)
 
 
-class ParseLooseVersionTest(unittest.TestCase):
-    """Test for certbot.util.parse_loose_version.
+class LooseVersionTest(unittest.TestCase):
+    """Test for certbot.util.LooseVersion.
 
     These tests are based on the original tests for
     distutils.version.LooseVersion at
@@ -643,26 +643,41 @@ class ParseLooseVersionTest(unittest.TestCase):
 
     @classmethod
     def _call(cls, *args, **kwargs):
-        from certbot.util import parse_loose_version
-        return parse_loose_version(*args, **kwargs)
+        from certbot.util import LooseVersion
+        return LooseVersion(*args, **kwargs)
 
     def test_less_than(self):
         comparisons = (('1.5.1', '1.5.2b2'),
             ('3.4j', '1996.07.12'),
             ('2g6', '11g'),
             ('0.960923', '2.2beta29'),
-            ('1.13++', '5.5.kw'))
+            ('1.13++', '5.5.kw'),
+            ('2.0', '2.0.1'),
+            ('a', 'b'))
         for v1, v2 in comparisons:
             assert self._call(v1) < self._call(v2)
 
     def test_equal(self):
-        assert self._call('8.02') == self._call('8.02')
+        comparisons = (('8.02', '8.02'),
+            ('1a', '1a'),
+            ('2.0', '2.0.0'))
+        for v1, v2 in comparisons:
+            assert self._call(v1) == self._call(v2)
 
     def test_greater_than(self):
         comparisons = (('161', '3.10a'),
             ('3.2.pl0', '3.1.1.6'))
         for v1, v2 in comparisons:
             assert self._call(v1) > self._call(v2)
+
+    def test_incomparible(self):
+        comparisons = (('bookworm/sid', '9'),
+            ('1a', '1.0'))
+        for v1, v2 in comparisons:
+            assert not self._call(v1) < self._call(v2)
+            assert not self._call(v1) > self._call(v2)
+            assert not self._call(v1) == self._call(v2)
+            assert self._call(v1) != self._call(v2)
 
 
 if __name__ == "__main__":